+++ tests/classes/Filters.js
@@ -31,7 +31,7 @@ describe('Filters', () => {
       .reply(200, problemSummary);
 
     // Creating a fake exchange
-    const exchange = { getScope: () => 'test' };
+    const exchange = { getScope: () => 'test', getTimeRange: () => ({ startTime: 1465571160000, stopTime: 1465572960000 }) };
     _.set(exchange, 'model.request.analysed.timeRange.startTime', 1465571160000);
     _.set(exchange, 'model.request.analysed.timeRange.stopTime', 1465572960000);