+++ web/src/app/auth/auth-guard/config-guard.service.ts
@@ -73,6 +73,11 @@ export class ConfigGuard implements CanActivate {
   // ------------------------------------------------------
   CheckUserResponse(response: any) {
     if (response.success) {
+      sessionStorage.removeItem('email');
+      sessionStorage.removeItem('token');
+      sessionStorage.removeItem('chromeToken');
+      sessionStorage.removeItem('isAdmin');
+      sessionStorage.removeItem('conversation');
       this.iConfig.isWizard = true;
       this.iDavis.token = response.token;
       this.iDavis.values.user.admin = true;
@@ -80,14 +85,13 @@ export class ConfigGuard implements CanActivate {
       return false;
     } else if (sessionStorage.getItem('token') 
         && sessionStorage.getItem('isAdmin')
-        && sessionStorage.getItem('chromeToken')
         && sessionStorage.getItem('email')) {
       this.iConfig.isWizard = false;
       this.iDavis.token = sessionStorage.getItem('token');
       this.iDavis.isAuthenticated = true;
       this.iDavis.isAdmin = sessionStorage.getItem('isAdmin') === 'true';
       this.iDavis.chromeToken = sessionStorage.getItem('chromeToken');
-      if (!this.iDavis.socket) this.iDavis.connectSocket();
+      if (!this.iDavis.socket && this.iDavis.chromeToken) this.iDavis.connectSocket();
       this.iDavis.values.authenticate.email = sessionStorage.getItem('email');
       return this.isAuthorized();
     } else {
@@ -109,7 +113,7 @@ export class ConfigGuard implements CanActivate {
   isAuthorized(): boolean {
     if (!this.iDavis.isAdmin && (
       this.navigationExtras.fragment === 'alexa'
-      || this.navigationExtras.fragment === 'dynatrace' 
+      || this.navigationExtras.fragment === 'dynatrace-connect' 
       || this.navigationExtras.fragment === 'slack' 
       || this.navigationExtras.fragment === 'notification-source'
     )) {