+++ web/src/app/auth/auth-guard/davis-guard.service.ts
@@ -61,6 +61,11 @@ export class DavisGuard implements CanActivate {
   // ------------------------------------------------------
   CheckUserResponse(response: any) {
     if (response.success) {
+      sessionStorage.removeItem('email');
+      sessionStorage.removeItem('token');
+      sessionStorage.removeItem('chromeToken');
+      sessionStorage.removeItem('isAdmin');
+      sessionStorage.removeItem('conversation');
       this.iConfig.isWizard = true;
       this.iDavis.token = response.token;
       this.iDavis.values.user.admin = true;
@@ -68,14 +73,13 @@ export class DavisGuard implements CanActivate {
       return false;
     } else if (sessionStorage.getItem('token') 
         && sessionStorage.getItem('isAdmin')
-        && sessionStorage.getItem('chromeToken')
         && sessionStorage.getItem('email')) {
       this.iConfig.isWizard = false;
       this.iDavis.token = sessionStorage.getItem('token'); 
       this.iDavis.isAuthenticated = true;
       this.iDavis.isAdmin = sessionStorage.getItem('isAdmin') === 'true';
       this.iDavis.chromeToken = sessionStorage.getItem('chromeToken');
-      if (!this.iDavis.socket) this.iDavis.connectSocket();
+      if (!this.iDavis.socket && this.iDavis.chromeToken) this.iDavis.connectSocket();
       this.iDavis.values.authenticate.email = sessionStorage.getItem('email');
       return true;
     } else {