+++ web/src/app/shared/config/config-dynatrace-connect/config-dynatrace-connect.component.ts
@@ -10,10 +10,10 @@ import { DavisService }   from '../../davis.service';
 import * as _ from "lodash";
 
 @Component({
-  selector: 'config-dynatrace',
-  templateUrl: './config-dynatrace.component.html',
+  selector: 'config-dynatrace-connect',
+  templateUrl: './config-dynatrace-connect.component.html',
 })
-export class ConfigDynatraceComponent implements OnInit, AfterViewInit {
+export class ConfigDynatraceConnectComponent implements OnInit, AfterViewInit {
 
   @ViewChild('url') url: ElementRef;
   
@@ -39,7 +39,7 @@ export class ConfigDynatraceComponent implements OnInit, AfterViewInit {
     }
     this.iConfig.connectDynatrace()
       .then(response => {
-        if (!response.success) { 
+        if (!response.success) {
           this.resetSubmitButton(); 
           throw new Error(response.message); 
         }
@@ -51,15 +51,17 @@ export class ConfigDynatraceComponent implements OnInit, AfterViewInit {
           throw new Error(response.message); 
         }
         
-        this.iConfig.status['dynatrace'].success = true;
+        this.iConfig.status['dynatrace-connect'].success = true;
         if (this.iConfig.isWizard) {
           this.iConfig.selectView('user');
         } else {
+          this.iConfig.values.original.dynatrace = _.cloneDeep(this.iConfig.values.dynatrace);
           this.resetSubmitButton();
+          this.validate();
         }
       })
       .catch(err => {
-        this.iConfig.displayError(err, 'dynatrace');
+        this.iConfig.displayError(err, 'dynatrace-connect');
       });
   }
 
@@ -77,8 +79,8 @@ export class ConfigDynatraceComponent implements OnInit, AfterViewInit {
   
   ngAfterViewInit() {
     if (window.location.protocol === 'http:' && window.location.hostname !== 'localhost' && window.location.hostname !== '127.0.01') {
-      this.iConfig.status['dynatrace'].error = 'Warning, please note that a valid SSL certificate is required to use davis!';
-      this.iConfig.status['dynatrace'].success = false;
+      this.iConfig.status['dynatrace-connect'].error = 'Warning, please note that a valid SSL certificate is required to use davis!';
+      this.iConfig.status['dynatrace-connect'].success = false;
     } else {
       this.isSSL = true;
     }
@@ -92,6 +94,18 @@ export class ConfigDynatraceComponent implements OnInit, AfterViewInit {
         this.validate();
       }, 1000);
     }
-    this.validate();
+
+    if (!this.iConfig.isWizard) {
+      this.iConfig.getDynatrace()
+        .then(response => {
+          if (!response.success) throw new Error(response.message);
+          this.iConfig.values.dynatrace = response.dynatrace;
+          this.iConfig.values.original.dynatrace = _.cloneDeep(this.iConfig.values.dynatrace);
+          this.validate();
+        })
+        .catch(err => {
+          this.iConfig.displayError(err, null);
+        });
+    }
   }
 }