+++ CHANGELOG.md
@@ -1,5 +1,31 @@
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/Dynatrace/davis-server/compare/v0.11.1...v0.12.0) (2017-04-06)
+
+
+### Bug Fixes
+
+* Remove sessionStorage items when isWizard is true in Angular guards ([69b9316](https://github.com/Dynatrace/davis-server/commit/69b9316))
+
+
+### Features
+
+* Add push notification support to web UI and Chrome extension [#182](https://github.com/Dynatrace/davis-server/issues/182). feat: Add support for message history in web ui [#161](https://github.com/Dynatrace/davis-server/issues/161) ([44aace4](https://github.com/Dynatrace/davis-server/commit/44aace4))
+* Added a cancel intent ([4fb91c2](https://github.com/Dynatrace/davis-server/commit/4fb91c2))
+* Create Open Dynatrace plugin [#245](https://github.com/Dynatrace/davis-server/issues/245) ([b270c61](https://github.com/Dynatrace/davis-server/commit/b270c61))
+* Image support in web UI cards [#238](https://github.com/Dynatrace/davis-server/issues/238) ([d986c65](https://github.com/Dynatrace/davis-server/commit/d986c65))
+* webChecks plugin [#248](https://github.com/Dynatrace/davis-server/issues/248) ([31622d9](https://github.com/Dynatrace/davis-server/commit/31622d9))
+* x-router-token authentication ([66229fb](https://github.com/Dynatrace/davis-server/commit/66229fb))
+
+
+
 <a name="0.11.1"></a>
 ## [0.11.1](https://github.com/Dynatrace/davis-server/compare/v0.11.0...v0.11.1) (2017-03-06)
+
+
+### Bug Fixes
+
+* [#247](https://github.com/Dynatrace/davis-server/issues/247) Multiple RTM connections ([63bb8c6](https://github.com/Dynatrace/davis-server/commit/63bb8c6))
+* Drop index on empty DB error ([ffb0b86](https://github.com/Dynatrace/davis-server/commit/ffb0b86))